+++ tests/testthat/test-produce_ci.R
@@ -1,0 +1,4 @@
+test_that("PA produces CI", {
+  mbsem_test_pa_ci(fit_pa, TRUE)
+  mbsem_test_pa_ci(fit_pa, FALSE)
+})